Remove ConstraintLayout
Using ConstraintLayout for such a simple application really is a waste. And it made my APK double in size, compared to the RelativeLayout version.

Additionally, remove test instrumentation (no need to test an app that just shows a layout), and enable ProGuard for release builds, just for kicks.

Also, update targetSdkVersion to 27, along with the support library version and compileSdkVersion.

Also, update the kotlin version so that it's current.